Run Release Cloud candidate tests with Jest
